Now does not write extra CRs.
Changed the output file to binary mode with no encoding.  I'm not sure why, but this writes utf-8 without double CRs (on Windows).  Not sure what it will do on UNIX.

Added documentation of the config file to README.

Changed version to 0.1.1.

Issue: the config file is currently not utf-8, so it might not handle non-english characters in titles.