Added function msg.
Now writes messages to stdout to try to make it work with a .cmd file.  Made no change.  The startup messages from a cmd file when the Python program is run, when the executable produced by py2exe is run, but not when the NSIS executable is run.  An alternative to the NSIS method might be to use the py2exe bundle_files=1 option, but that doesn't currently work on 64-bit (Python or machines?).

Added RequestExecutionLevel user to setup.nsi to keep UAC from prompting.