Added md2html.au3.
Messages never appear from the executable produced by setup.nsi.   A dev environment misconfiguration was causing the wrong executable to be run when testing.

Added md2html.au3, an AutoIt script which, when compiled into an EXE installs the necessary files in a temp directory, runs the program, and deletes the temp folder.  A /install command line switch causes it to prompt for a folder to install into.

Added build.cmd which runs Python to build the "real" executable and then Aut2Exe to make the "installer".

Some tests for "-" as input file still exist in md2html.py, but this is not supported.  They are left because they protect the config file code in the event reading from stdin were restored in the future.

setup.nsi is no longer used.  It is left in case it might be useful in the future.

Changed version to 1.0.3.